maptry/requirements.txt
ortools>=9.5.0
openrouteservice>=2.3.3
geopy>=2.3.0
folium>=0.14.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
